Infra: eliminate all compiler warnings and enable /WX
Fix 8 warnings across 3 codes that appeared in Level4 x64 builds:

  C4267 (×3) — size_t→int narrowing at three call sites:
    static_cast<int>(m_MaxDepth) in AIPerplex::GetMove()
    static_cast<int>(ply) in AIAgent::Search() and AIBasic::Search()

  C4100 (×4) — assert_tt_store params only live inside assert() macros,
    which are no-ops in Release/NDEBUG. Fix: [[maybe_unused]] on value,
    depth, bound, and phase in both declaration and definition.

  C4244 (×1) — std::tolower returns int; lambda assigned to char.
    Fix: static_cast<char>(std::tolower(c)) in StrHelper::toLower.

Enable /WX (treat-warnings-as-errors) for Debug|x64 and Release|x64 in
both StratChessEvolved.vcxproj and StratChessTests.vcxproj. Upgrade the
test project from Level3 to Level4 for consistent enforcement.

Also registers the four WIP source files (AIPerplexParallel, MoveOrdering,
OrigPosition, Position) in the VS project as ExcludedFromBuild, keeping
Solution Explorer in sync without adding them to the build.

Both solutions build clean with zero warnings in Debug and Release.
All 214 test assertions pass.

+++ StratEngine/AIAgent.cpp
@@ -80,7 +80,7 @@ int AIAgent::Search(_In_ size_t ply, _In_ int alpha, _In_ int beta, _Inout_ PVLi
 	const GameInfo& info = GetLastBoardInfo( ply );
 
 	// Test for 50 moves rule
-	if ( checkDraws( info, ply) )
+	if ( checkDraws( info, static_cast<int>(ply)) )
 		return GameValues::Draw;
 
 	// Er vi naaet til bunden af traeet - leaf nodes?
